Moved code from RelationalType to ObjectType
