+++ layers/Engine/packages/component-model/src/HelperServices/DataloadHelperService.php
@@ -10,6 +10,7 @@ use PoP\ComponentModel\Schema\FeedbackMessageStoreInterface;
 use PoP\ComponentModel\Schema\FieldQueryInterpreterInterface;
 use PoP\ComponentModel\Schema\SchemaHelpers;
 use PoP\ComponentModel\TypeResolvers\RelationalTypeResolverInterface;
+use PoP\ComponentModel\TypeResolvers\Union\UnionTypeResolverInterface;
 use PoP\Translation\TranslationAPIInterface;
 
 class DataloadHelperService implements DataloadHelperServiceInterface
@@ -22,19 +23,35 @@ class DataloadHelperService implements DataloadHelperServiceInterface
     ) {
     }
 
+    /**
+     * Accept RelationalTypeResolverInterface as param, instead of the more natural
+     * ObjectTypeResolverInterface, to make it easy within the application to check
+     * for this result without checking in advance what's the typeResolver
+     */
     public function getTypeResolverClassFromSubcomponentDataField(RelationalTypeResolverInterface $relationalTypeResolver, string $subcomponent_data_field): ?string
     {
+        /**
+         * Because the UnionTypeResolver doesn't know yet which TypeResolver will be used
+         * (that depends on each resultItem), it can't resolve this functionality
+         */
+        if ($relationalTypeResolver instanceof UnionTypeResolverInterface) {
+            return null;
+        }
+        // By now, the typeResolver must be ObjectType
+        /** @var ObjectTypeResolverInterface */
+        $objectTypeResolver = $relationalTypeResolver;
+
         // If this field doesn't have a typeResolver, show a schema error
         // But if there are no FieldResolvers, then skip adding an error here, since that error will have been added already
         // Otherwise, there will appear 2 error messages:
         // 1. No FieldResolver
         // 2. No FieldDefaultTypeDataLoader
-        $subcomponentFieldTypeResolverClass = $relationalTypeResolver->getFieldTypeResolverClass($subcomponent_data_field);
-        if (!SchemaHelpers::isRelationalFieldTypeResolverClass($subcomponentFieldTypeResolverClass) && $relationalTypeResolver->hasFieldResolversForField($subcomponent_data_field)) {
+        $subcomponentFieldTypeResolverClass = $objectTypeResolver->getFieldTypeResolverClass($subcomponent_data_field);
+        if (!SchemaHelpers::isRelationalFieldTypeResolverClass($subcomponentFieldTypeResolverClass) && $objectTypeResolver->hasFieldResolversForField($subcomponent_data_field)) {
             // If there is an alias, store the results under this. Otherwise, on the fieldName+fieldArgs
             $subcomponent_data_field_outputkey = $this->fieldQueryInterpreter->getFieldOutputKey($subcomponent_data_field);
             $this->feedbackMessageStore->addSchemaError(
-                $relationalTypeResolver->getTypeOutputName(),
+                $objectTypeResolver->getTypeOutputName(),
                 $subcomponent_data_field_outputkey,
                 sprintf(
                     $this->translationAPI->__('Field \'%s\' is not a connection', 'pop-component-model'),
